#25 feat(search): 상품 아이템 Composable 분리
- `SearchScreen.kt`에 있던 `ProductItem` 컴포저블을 `presentation/component/ProductItem.kt` 파일로 분리
- `SearchScreen.kt`에서는 분리된 컴포저블을 import하여 사용하도록 수정함
- `SearchScreen.kt`에 `@Preview` 추가
